+++ crates/ruma-client-api/src/keys/upload_keys.rs
@@ -15,7 +15,7 @@ pub mod v3 {
         encryption::{DeviceKeys, OneTimeKey},
         metadata,
         serde::Raw,
-        DeviceKeyAlgorithm, OwnedDeviceKeyId,
+        OneTimeKeyAlgorithm, OwnedOneTimeKeyId,
     };
 
     const METADATA: Metadata = metadata! {
@@ -40,11 +40,11 @@ pub mod v3 {
 
         /// One-time public keys for "pre-key" messages.
         #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-        pub one_time_keys: BTreeMap<OwnedDeviceKeyId, Raw<OneTimeKey>>,
+        pub one_time_keys: BTreeMap<OwnedOneTimeKeyId, Raw<OneTimeKey>>,
 
         /// Fallback public keys for "pre-key" messages.
         #[serde(default, skip_serializing_if = "BTreeMap::is_empty")]
-        pub fallback_keys: BTreeMap<OwnedDeviceKeyId, Raw<OneTimeKey>>,
+        pub fallback_keys: BTreeMap<OwnedOneTimeKeyId, Raw<OneTimeKey>>,
     }
 
     /// Response type for the `upload_keys` endpoint.
@@ -52,7 +52,7 @@ pub mod v3 {
     pub struct Response {
         /// For each key algorithm, the number of unclaimed one-time keys of that
         /// type currently held on the server for this device.
-        pub one_time_key_counts: BTreeMap<DeviceKeyAlgorithm, UInt>,
+        pub one_time_key_counts: BTreeMap<OneTimeKeyAlgorithm, UInt>,
     }
 
     impl Request {
@@ -64,7 +64,7 @@ pub mod v3 {
 
     impl Response {
         /// Creates a new `Response` with the given one time key counts.
-        pub fn new(one_time_key_counts: BTreeMap<DeviceKeyAlgorithm, UInt>) -> Self {
+        pub fn new(one_time_key_counts: BTreeMap<OneTimeKeyAlgorithm, UInt>) -> Self {
             Self { one_time_key_counts }
         }
     }